Move URL utils to application commons
